requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
boto3==1.34.144
    # via -r requirements.in
botocore==1.34.144
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
brotli==1.1.0
    # via yt-dlp
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
    #   yt-dlp
charset-normalizer==3.3.2
    # via requests
decorator==4.4.2
    # via moviepy
distro==1.9.0
    # via openai
et-xmlfile==1.1.0
    # via openpyxl
filelock==3.15.4
    # via -r requirements.in
fsspec==2024.6.1
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via
    #   -r requirements.in
    #   httpx
httpx==0.27.0
    # via
    #   -r requirements.in
    #   openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
imageio==2.34.2
    # via moviepy
imageio-ffmpeg==0.5.1
    # via moviepy
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
moviepy==1.0.3
    # via -r requirements.in
mutagen==1.47.0
    # via
    #   -r requirements.in
    #   yt-dlp
numpy==2.0.0
    # via
    #   imageio
    #   moviepy
openai==1.35.9
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
pillow==10.4.0
    # via imageio
pinecone-client==4.1.2
    # via -r requirements.in
pinecone-plugin-interface==0.0.7
    # via
    #   -r requirements.in
    #   pinecone-client
proglog==0.1.10
    # via moviepy
pycryptodomex==3.20.0
    # via yt-dlp
pydantic==2.8.2
    # via openai
pydantic-core==2.20.1
    # via pydantic
pydub==0.25.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   moviepy
    #   yt-dlp
s3transfer==0.10.2
    # via
    #   -r requirements.in
    #   boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.4
    # via
    #   -r requirements.in
    #   moviepy
    #   openai
    #   pinecone-client
    #   proglog
typing-extensions==4.12.2
    # via
    #   openai
    #   pinecone-client
    #   pydantic
    #   pydantic-core
urllib3==2.2.2
    # via
    #   -r requirements.in
    #   botocore
    #   pinecone-client
    #   requests
    #   yt-dlp
websockets==12.0
    # via yt-dlp
yt-dlp==2024.7.9
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
